[OPIK-881] add plugin and update using playground tab
